Rename NBProject to match folder name
