+++ nbproject/project.properties
@@ -4,7 +4,7 @@ annotation.processing.processor.options=
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
-application.title=TutorialHelper
+application.title=EasyScreenshotHelper
 application.vendor=jakob
 build.classes.dir=${build.dir}/classes
 build.classes.excludes=**/*.java,**/*.form
@@ -26,7 +26,7 @@ debug.test.classpath=\
     ${run.test.classpath}
 # This directory is removed when the project is cleaned:
 dist.dir=dist
-dist.jar=${dist.dir}/TutorialHelper.jar
+dist.jar=${dist.dir}/EasyScreenshotHelper.jar
 dist.javadoc.dir=${dist.dir}/javadoc
 endorsed.classpath=
 excludes=
