+++ nbproject/project.xml
@@ -13,7 +13,7 @@
             </extension>
         </buildExtensions>
         <data xmlns="http://www.netbeans.org/ns/j2se-project/3">
-            <name>TutorialHelper</name>
+            <name>EasyScreenshotHelper</name>
             <source-roots>
                 <root id="src.dir"/>
             </source-roots>
