+++ src/easyscreenshothelper/GlobalKeyListener.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package tutorialhelper;
+package easyscreenshothelper;
 
 import java.util.Observable;
 import org.jnativehook.keyboard.NativeKeyEvent;
